requirements.txt
Adafruit-Blinka==6.15.0
adafruit-circuitpython-ads1x15==2.2.9
adafruit-circuitpython-busdevice==5.1.0
adafruit-circuitpython-dht==3.6.3
Adafruit-PlatformDetect==3.17.1
Adafruit-PureIO==1.1.9
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
paho-mqtt==1.6.1
pyftdi==0.53.3
pyserial==3.5
pyusb==1.2.1
requests==2.26.0
rpi-ws281x==4.3.1
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
tinydb==4.6.1
typing-extensions==4.1.1
urllib3==1.26.7
PyYAML~=6.0